scripts/requirements.txt
# scripts/requirements.txt
#
# Install with:
# pip install -r scripts/requirements.txt

transformers
torch
sacremoses
requests
PyYAML
beautifulsoup4
google-search-results
